[refactor] refactor server, decoupling server, tokenizer, model, backend, config, and some code format
