+++ src/models/baichuan/baichuan_tokenizer.h
@@ -1,15 +1,33 @@
-#ifndef __PPL_LLM_INTERNLM_TOKENIZER_H__
-#define __PPL_LLM_INTERNLM_TOKENIZER_H__
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+#ifndef __PPL_LLM_BAICHUAN_TOKENIZER_H__
+#define __PPL_LLM_BAICHUAN_TOKENIZER_H__
 
 #include "utils/tokenizer.h"
+
 #include "ppl/nn/common/logger.h"
-#include <sentencepiece_processor.h>
+#include "sentencepiece_processor.h"
 
 #include <string_view>
 
-namespace ppl { namespace llm {
+namespace ppl { namespace llm { namespace baichuan {
 
-class InternLmTokenizer final : public utils::Tokenizer {
+class BaiChuanTokenizer final : public utils::Tokenizer {
 public:
     ppl::common::RetCode Init(const std::string& path) {
         auto tokenizer_status = sp_processor_.Load(path);
@@ -23,7 +41,6 @@ public:
     }
 
     void Encode(const char* prompt, uint32_t len, std::vector<int>* token_ids) const override {
-        token_ids->push_back(sp_processor_.bos_id()); // preprocess
         sp_processor_.Encode(std::string_view(prompt, len), token_ids);
     }
 
@@ -47,6 +64,6 @@ private:
     sentencepiece::SentencePieceProcessor sp_processor_;
 };
 
-}} // namespace ppl::llm
+}}} // namespace ppl::llm::baichuan
 
 #endif
